+++ src/App.js
@@ -1,70 +1,17 @@
-import React, { useEffect } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import ReactLoading from 'react-loading';
-import {
-  selectAllCovidData,
-  getCovidStatus,
-  getCovidError,
-  fetchCovidData,
-  getPageTitle,
-  getSearchResults,
-} from './features/covid/covidSlice';
-import Categories from './components/Categories';
-import Navbar from './components/navbar';
+import React from 'react';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import CovidIndex from './features/covid/covid';
+import CovidDetailsPage from './features/covid/covidDetails';
 
-const App = () => {
-  const dispatch = useDispatch();
-  const covidData = useSelector(selectAllCovidData);
-  const loadingStatus = useSelector(getCovidStatus);
-  const error = useSelector(getCovidError);
-  const pageTitle = useSelector(getPageTitle);
-  const searchResults = useSelector(getSearchResults);
-
-  useEffect(() => {
-    if (loadingStatus === 'idle') {
-      dispatch(fetchCovidData());
-    }
-  }, [loadingStatus, dispatch]);
-
-  let newsContent = '';
-  if (loadingStatus === 'loading') {
-    newsContent = (
-      <div
-        style={{
-          margin: '2rem auto',
-        }}
-      >
-        <ReactLoading type="spin" color="blue" height={100} width={100} />
-      </div>
-    );
-  } else if (loadingStatus === 'succeeded') {
-    const dataToRender = searchResults.length > 0 ? searchResults : covidData;
-    newsContent = dataToRender.map((data) => (
-      <div key={data.item_id}>
-        <Categories
-          abbreviation={data.abbreviation ? data.abbreviation : 'GH'}
-          country={data.item_id}
-          confirmed={data.confirmed}
-          deaths={data.deaths}
-          covidData={covidData}
-        />
-      </div>
-    ));
-  } else if (loadingStatus === 'failed') {
-    newsContent = <h2>{error}</h2>;
-  }
-
-  return (
-    <div>
-      <Navbar />
-      <main className="container py-5">
-        <h1>{pageTitle}</h1>
-        <div className="row row-cols-1 row-cols-sm-2 row-cols-md-3 g-3">
-          {newsContent}
-        </div>
-      </main>
-    </div>
-  );
-};
+const App = () => (
+  <div>
+    <BrowserRouter>
+      <Routes>
+        <Route exact path="/" element={<CovidIndex />} />
+        <Route exact path="/country/:name" element={<CovidDetailsPage />} />
+      </Routes>
+    </BrowserRouter>
+  </div>
+);
 
 export default App;
